moved isOnline() back into ConnectionManager, for config and logger
