+++ app/helpers/httpHelper.js
@@ -10,28 +10,7 @@ class HTTPHelper {
 			processRequest(url, res, rej);
 		});
 	}
-
-	/**
-	 * @param {number} timeout 
-	 * @param {number} retries 
-	 * @returns 
-	 */
-	async isOnline(timeout, retries) {
-		var resolved = false;
-		for (var i = 1; i <= retries && !resolved; i++) {
-			// Not using net.isOnline(), because it's too optimistic, it returns
-			// true before we can actually issue successful https requests.
-			resolved = await isOnlineInternal();
-			if (!resolved) await sleep(timeout);
-		}
-		return resolved;
-	}
 }
-
-function sleep(timeout) {
-	return new Promise(r => setTimeout(r, timeout));
-}
-
 
 function removeLeadingSlash(url) {
 	return (url[0] == '/') ? url = url.substr(1) : url;
@@ -64,22 +43,6 @@ function processRequest(url, resolve, reject) {
 	});
 
 	request.end();
-}
-
-function isOnlineInternal() {
-	return new Promise((resolve) => {
-		var req = net.request({
-			url: 'https://teams.microsoft.com',
-			method: 'HEAD'
-		});
-		req.on('response', () => {
-			resolve(true);
-		});
-		req.on('error', () => {
-			resolve(false);
-		});
-		req.end();
-	});
 }
 
 module.exports = new HTTPHelper();
